requirements.txt
openai>=1.0.0
pandas>=1.5.0
scikit-learn>=1.3.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.64.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
flask>=2.0.0
